[REF] web,*: calendar: regroup utils in a single module
*hr_homeworking_calendar

Followup of task-4510549
