+++ addons/web/static/src/views/calendar/calendar_filter_section/calendar_filter_section.js
@@ -3,7 +3,7 @@ import { AutoComplete } from "@web/core/autocomplete/autocomplete";
 import { Transition } from "@web/core/transition";
 import { useOwnedDialogs, useService } from "@web/core/utils/hooks";
 import { SelectCreateDialog } from "@web/views/view_dialogs/select_create_dialog";
-import { getColor } from "../colors";
+import { getColor } from "../utils";
 import { Component, useState } from "@odoo/owl";
 
 let nextId = 1;
