+++ addons/web/static/src/views/calendar/mobile_filter_panel/calendar_mobile_filter_panel.js
@@ -1,5 +1,5 @@
 import { Component } from "@odoo/owl";
-import { getColor } from "../colors";
+import { getColor } from "../utils";
 
 export class CalendarMobileFilterPanel extends Component {
     static components = {};
